Refactor to ensure app runs after merging with main
