+++ android/app/src/main/AndroidManifest.xml
@@ -1,23 +1,16 @@
-<<<<<<< HEAD
-<manifest xmlns:android="http://schemas.android.com/apk/res/android">
-=======
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.yourapp.tindahanap">
+    package="com.example.tindahanap">
     
     <!-- Add permissions for location access -->
     <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION"/>
     <uses-permission android:name="android.permission.ACCESS_COARSE_LOCATION"/>
 
->>>>>>> master
     <application
         android:label="tindahanap"
         android:name="${applicationName}"
         android:icon="@mipmap/ic_launcher">
-<<<<<<< HEAD
-=======
         
         <!-- Your main activity -->
->>>>>>> master
         <activity
             android:name=".MainActivity"
             android:exported="true"
@@ -27,45 +20,23 @@
             android:configChanges="orientation|keyboardHidden|keyboard|screenSize|smallestScreenSize|locale|layoutDirection|fontScale|screenLayout|density|uiMode"
             android:hardwareAccelerated="true"
             android:windowSoftInputMode="adjustResize">
-<<<<<<< HEAD
-            <!-- Specifies an Android theme to apply to this Activity as soon as
-                 the Android process has started. This theme is visible to the user
-                 while the Flutter UI initializes. After that, this theme continues
-                 to determine the Window background behind the Flutter UI. -->
-            <meta-data
-              android:name="io.flutter.embedding.android.NormalTheme"
-              android:resource="@style/NormalTheme"
-              />
-=======
             
             <meta-data
               android:name="io.flutter.embedding.android.NormalTheme"
               android:resource="@style/NormalTheme"
             />
             
->>>>>>> master
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
-<<<<<<< HEAD
-=======
         
->>>>>>> master
         <!-- Don't delete the meta-data below.
              This is used by the Flutter tool to generate GeneratedPluginRegistrant.java -->
         <meta-data
             android:name="flutterEmbedding"
             android:value="2" />
-<<<<<<< HEAD
-    </application>
-    <!-- Required to query activities that can process text, see:
-         https://developer.android.com/training/package-visibility and
-         https://developer.android.com/reference/android/content/Intent#ACTION_PROCESS_TEXT.
-
-         In particular, this is used by the Flutter engine in io.flutter.plugin.text.ProcessTextPlugin. -->
-=======
 
         <!-- Google Maps API key -->
         <meta-data
@@ -75,15 +46,10 @@
     </application>
 
     <!-- Required to query activities that can process text -->
->>>>>>> master
     <queries>
         <intent>
             <action android:name="android.intent.action.PROCESS_TEXT"/>
             <data android:mimeType="text/plain"/>
         </intent>
     </queries>
-<<<<<<< HEAD
-=======
-
->>>>>>> master
 </manifest>
