+++ lib/main.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:firebase_core/firebase_core.dart';
-import 'package:flutter/material.dart';
-import 'ui/theme/app_theme.dart';
-import 'ui/screens/login_screen.dart';
+import 'theme/app_theme.dart';
+import 'screens/login_screen.dart';
 import 'screens/homepage.dart';
 
 void main() async {
@@ -10,6 +9,7 @@ void main() async {
   
   // Firebase initialization with the config for Flutter
   await Firebase.initializeApp(
+    name: 'tindahanap-app',
     options: FirebaseOptions(
       apiKey: "AIzaSyAY-2R0EX1vY1T5UTSJSrVLV-whQSmjU1w",
       appId: "1:342791532648:web:85714f4e0b1c238cb6cbdd",
