--- lib/screens/login_page.dart
@@ -1,131 +1,0 @@
-import 'package:flutter/material.dart';
-import '../services/auth_service.dart';
-import 'homepage.dart';
-import 'signup_page.dart';
-
-class LoginPage extends StatefulWidget {
-  @override
-  _LoginPageState createState() => _LoginPageState();
-}
-
-class _LoginPageState extends State<LoginPage> {
-  final _authService = AuthService();
-  final _usernameController = TextEditingController();
-  final _passwordController = TextEditingController();
-
-  void login() async {
-    bool success = await _authService.login(
-      _usernameController.text,
-      _passwordController.text,
-    );
-    if (success) {
-      Navigator.pushReplacement(
-        context,
-        MaterialPageRoute(builder: (_) => HomePage()),
-      );
-    } else {
-      ScaffoldMessenger.of(context).showSnackBar(
-        SnackBar(content: Text("Invalid Username or Password")),
-      );
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Colors.white,
-      body: Center(
-        child: Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 30),
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: [
-              Container(
-                decoration: BoxDecoration(
-                  color: Color.fromARGB(255, 162, 126, 223),
-                  borderRadius: BorderRadius.circular(20),
-                ),
-                padding: EdgeInsets.all(20),
-                child: Icon(
-                  Icons.person,
-                  size: 80,
-                  color: Color.fromARGB(255, 78, 14, 190),
-                ),
-              ),
-              SizedBox(height: 30),
-              
-              TextField(
-                controller: _usernameController,
-                decoration: InputDecoration(
-                  prefixIcon: Icon(Icons.email, color: Color.fromARGB(255, 78, 14, 190)),
-                  labelText: "Email Address",
-                  border: OutlineInputBorder(
-                    borderRadius: BorderRadius.circular(10),
-                  ),
-                ),
-              ),
-              SizedBox(height: 20),
-              
-              TextField(
-                controller: _passwordController,
-                obscureText: true,
-                decoration: InputDecoration(
-                  prefixIcon: Icon(Icons.lock, color: Color.fromARGB(255, 78, 14, 190)),
-                  labelText: "Password",
-                  border: OutlineInputBorder(
-                    borderRadius: BorderRadius.circular(10),
-                  ),
-                ),
-              ),
-              SizedBox(height: 10),
-              
-              Align(
-                alignment: Alignment.centerRight,
-                child: TextButton(
-                  onPressed: () {},
-                  child: Text("Forgot Password?", style: TextStyle(color: Color.fromARGB(255, 78, 14, 190))),
-                ),
-              ),
-              SizedBox(height: 10),
-              
-              SizedBox(
-                width: double.infinity,
-                height: 50,
-                child: ElevatedButton(
-                  onPressed: login,
-                  style: ElevatedButton.styleFrom(
-                    backgroundColor: Color.fromARGB(255, 78, 14, 190),
-                    shape: RoundedRectangleBorder(
-                      borderRadius: BorderRadius.circular(10),
-                    ),
-                  ),
-                  child: Text("Log In", style: TextStyle(fontSize: 18, color: Colors.white)),
-                ),
-              ),
-              SizedBox(height: 20),
-              
-              Row(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: [
-                  Text("Don't have an account? "),
-                  GestureDetector(
-                    onTap: () {
-                      Navigator.push(
-                        context,
-                        MaterialPageRoute(builder: (context) => SignupPage()),
-                      );
-                    },
-                    child: Text(
-                      "Sign Up",
-                      style: TextStyle(fontWeight: FontWeight.bold, color: Color.fromARGB(255, 78, 14, 190)),
-                    ),
-                  ),
-                ],
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
