+++ lib/screens/login_screen.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:email_validator/email_validator.dart';
-import '../../auth_service.dart';
-import '../../homepage.dart';
-import '../../signup_page.dart';
+import '../services/auth_service.dart';
+import 'homepage.dart';
+import 'signup_page.dart';
 
 class LoginPage extends StatefulWidget {
   const LoginPage({super.key});
