+++ pubspec.yaml
@@ -42,8 +42,8 @@ dependencies:
   firebase_data_connect: ^0.1.4
   google_maps_flutter: ^2.5.0
   http: ^1.2.1
-  geolocator: ^8.0.0
-  permission_handler: ^10.0.0
+  geolocator: ^13.0.4
+  permission_handler: ^11.4.0
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
@@ -63,7 +63,7 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^4.0.0
+  flutter_lints: ^5.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
